refactor(chat): extract impersonation + per-speaker realism orchestration
Continues the chat_service.dart decomposition. Three more cohesive clusters
move out of the coordinator into `part`/extension files (verbatim, zero
behaviour change):

- impersonateUser (generate a message in the user's voice)
  -> chat_service_impersonate.dart
- the per-speaker realism eval + the load/save impersonation "dance"
  (_evaluateRealismForUpcomingSpeaker, _loadGroupRealismIntoScalars,
  _saveScalarsIntoGroupRealism)
  -> chat_service_realism_dance.dart
- per-speaker objective focus/seed + realism-state injection/restore +
  post-gen needs checks + mood decay
  -> chat_service_speaker_objectives.dart

Two cluster-local static references (defaultGroupSystemPrompt,
observerModeSystemPrompt, defaultApiSystemPrompt, defaultKoboldSystemPrompt)
had to be qualified with `ChatService.` once moved into an extension, since
extension bodies can't reach the host type's static members unqualified.

getActiveObjectivesFor was deliberately kept in the class body rather than
moved with the rest of the objectives cluster: FakeChatService overrides it in
the golden tests, and extension members are statically dispatched (an
`@override` on an extension method doesn't bind). A comment now records why it
must stay in-class so a future extractor doesn't repeat the mistake.

Net effect: chat_service.dart drops 5,603 -> 4,742 lines (-861). No public API,
persistence shape, or realism/needs behaviour changes.

flutter analyze: clean. Full suite (1348 unit + 94 golden) + the 1:1<->group
realism parity gate: green.
