+++ lib/services/chat_service.dart
@@ -97,6 +97,9 @@ part 'chat/chat_service_chat_entry.dart';
 part 'chat/chat_service_group_entry.dart';
 part 'chat/chat_service_session_state.dart';
 part 'chat/chat_service_session_load.dart';
+part 'chat/chat_service_realism_dance.dart';
+part 'chat/chat_service_speaker_objectives.dart';
+part 'chat/chat_service_impersonate.dart';
 part 'chat/chat_service_session_manage.dart';
 part 'chat/chat_service_generation.dart';
 
@@ -3286,376 +3289,6 @@ class ChatService extends ChangeNotifier {
     }
   }
 
-  Future<void> impersonateUser({
-    String prefix = '',
-    required Function(String accumulated) onToken,
-  }) async {
-    if ((_activeCharacter == null && _activeGroup == null) ||
-        _isGenerating ||
-        _guestBusy) {
-      return;
-    }
-
-    _isGenerating = true;
-    _cancelRequested = false;
-    notifyListeners();
-
-    try {
-      final userName = _userPersonaService.persona.name;
-
-      // Determine the speaking character (needed for prompt construction)
-      CharacterCard speakingCharacter;
-      if (_activeGroup != null) {
-        speakingCharacter = _groupCharacters.first;
-      } else {
-        speakingCharacter = _activeCharacter!;
-      }
-
-      // Build prompt the same way _generateResponse does
-      // Path B clean hierarchy (same as the main generation path)
-      String systemPrompt;
-      if (_activeGroup != null && _activeGroup!.systemPrompt.isNotEmpty) {
-        systemPrompt = _activeGroup!.systemPrompt;
-      } else if (_activeGroup != null) {
-        systemPrompt = _observerMode
-            ? observerModeSystemPrompt
-            : defaultGroupSystemPrompt;
-      } else if (speakingCharacter.systemPrompt.isNotEmpty) {
-        systemPrompt = speakingCharacter.systemPrompt;
-      } else if (_storageService.generationSettings.systemPrompt.isNotEmpty) {
-        systemPrompt = _storageService.generationSettings.systemPrompt;
-      } else {
-        final isApi = _llmProvider != null && !_llmProvider!.isLocal;
-        systemPrompt = isApi
-            ? defaultApiSystemPrompt
-            : defaultKoboldSystemPrompt;
-      }
-
-      if (_activeGroup != null) {
-        final groupCharPrompt = getSystemPromptForGroupCharacter(
-          speakingCharacter,
-        ).trim();
-        if (groupCharPrompt.isNotEmpty) {
-          systemPrompt +=
-              '\n\n[Group-specific instructions for ${speakingCharacter.name}]\n$groupCharPrompt';
-        } else if (speakingCharacter.systemPrompt.isNotEmpty) {
-          systemPrompt +=
-              '\n\n[Specific instructions for ${speakingCharacter.name}]\n${speakingCharacter.systemPrompt.trim()}';
-        }
-      }
-
-      // Lorebook (group + per-character, respecting inherit flag and group worlds)
-      String loreContent = '';
-      final activeLoreStrings = <String>{}; // Set for deduplication
-
-      final inherit = _activeGroup?.inheritCharacterLorebooks ?? true;
-
-      // Group-level lorebook (highest priority when present)
-      if (_activeGroup != null && _activeGroup!.groupLorebook.isNotEmpty) {
-        try {
-          final json = jsonDecode(_activeGroup!.groupLorebook);
-          final gl = Lorebook.fromJson(json as Map<String, dynamic>);
-          final active = gl.entries.where(
-            (e) => e.enabled && (e.isTriggered || e.constant),
-          );
-          activeLoreStrings.addAll(active.map((e) => e.content));
-        } catch (_) {}
-      }
-
-      // Group-level worlds (always included if attached to the group)
-      if (_activeGroup != null) {
-        for (final wid in _activeGroup!.worldIds) {
-          final world = _worldRepository.worlds
-              .where((w) => w.name == wid)
-              .firstOrNull;
-          if (world == null) continue;
-          final active = world.lorebook.entries.where(
-            (e) => e.enabled && (e.isTriggered || e.constant),
-          );
-          activeLoreStrings.addAll(active.map((e) => e.content));
-        }
-      }
-
-      // Per-character lore and their worlds (only if inherit is true or no group)
-      if (inherit || _activeGroup == null) {
-        final loreCharacters = _activeGroup != null
-            ? _groupCharacters
-            : (_activeCharacter != null
-                  ? [_activeCharacter!]
-                  : <CharacterCard>[]);
-        for (final ch in loreCharacters) {
-          if (ch.lorebook != null) {
-            final activeEntries = ch.lorebook!.entries.where(
-              (e) => e.enabled && (e.isTriggered || e.constant),
-            );
-            activeLoreStrings.addAll(activeEntries.map((e) => e.content));
-          }
-          for (final worldName in ch.worldNames) {
-            final world = _worldRepository.worlds
-                .where((w) => w.name == worldName)
-                .firstOrNull;
-            if (world == null) continue;
-            final activeWorldEntries = world.lorebook.entries.where(
-              (e) => e.enabled && (e.isTriggered || e.constant),
-            );
-            activeLoreStrings.addAll(activeWorldEntries.map((e) => e.content));
-          }
-        }
-      }
-
-      if (activeLoreStrings.isNotEmpty) {
-        loreContent = "Context Info:\n${activeLoreStrings.join('\n')}\n";
-      }
-
-      // Persona & scenario
-      // Use evolved versions if character evolution is enabled and available
-      String personaBlock;
-      if (_activeGroup != null) {
-        final personas = _groupCharacters
-            .map(
-              (ch) =>
-                  "${ch.name}'s Persona: ${_macroResolver.resolve(
-                    _getEffectivePersonality(ch),
-                    MacroContext(userName: userName, characterName: ch.name),
-                    section: 'persona',
-                  )}",
-            )
-            .toList();
-        personaBlock = personas.join('\n');
-      } else {
-        personaBlock =
-            "${speakingCharacter.name}'s Persona: ${_macroResolver.resolve(
-              _getEffectivePersonality(speakingCharacter),
-              MacroContext(userName: userName, characterName: speakingCharacter.name),
-              section: 'persona',
-            )}";
-      }
-
-      // User persona — inject user's self-description + learned facts
-      final userPersonaBlock = await _buildUserPersonaBlock(userName);
-
-      String rawScenario = '';
-      if (_activeGroup != null && _activeGroup!.scenario.isNotEmpty) {
-        rawScenario = _activeGroup!.scenario;
-      } else {
-        final scenarioChar = _activeGroup != null
-            ? _groupCharacters.first
-            : speakingCharacter;
-        rawScenario = _getEffectiveScenario(scenarioChar);
-      }
-      String scenario = rawScenario;
-
-      String history = _buildChatHistory();
-
-      // Suffix: user name + any partial text the user typed
-      String suffix = "\n$userName:";
-      if (prefix.isNotEmpty) {
-        suffix = "$suffix $prefix";
-      }
-
-      String mesExampleBlock = '';
-      if (_activeGroup != null) {
-        final examples = _groupCharacters
-            .where((ch) => ch.mesExample.isNotEmpty)
-            .map(
-              (ch) => _macroResolver.resolve(
-                ch.mesExample,
-                MacroContext(userName: userName, characterName: ch.name),
-                section: 'mesExample',
-              ),
-            )
-            .toList();
-        if (examples.isNotEmpty) {
-          mesExampleBlock = '${examples.join('\n')}\n';
-        }
-      } else if (speakingCharacter.mesExample.isNotEmpty) {
-        mesExampleBlock = '${speakingCharacter.mesExample}\n';
-      }
-
-      String postHistoryBlock = '';
-      if (_activeGroup == null &&
-          speakingCharacter.postHistoryInstructions.isNotEmpty) {
-        postHistoryBlock = '${speakingCharacter.postHistoryInstructions}\n';
-      }
-
-      String authorNoteBlock = '';
-      if (_authorNote.isNotEmpty) {
-        authorNoteBlock = _buildAuthorNoteBlock();
-      }
-
-      // ── Macro resolution pass ──
-      final macroCtx = MacroContext(
-        userName: userName,
-        characterName: speakingCharacter.name,
-        summaryMaxWords: _storageService.memorySettings.summaryMaxWords,
-        chatId: _currentSessionId,
-        characterId: speakingCharacter.dbId,
-      );
-      systemPrompt = _macroResolver.resolve(
-        systemPrompt,
-        macroCtx,
-        section: 'systemPrompt',
-      );
-      if (loreContent.isNotEmpty) {
-        loreContent = _macroResolver.resolve(
-          loreContent,
-          macroCtx,
-          section: 'lore',
-        );
-      }
-      scenario = _macroResolver.resolve(
-        scenario,
-        macroCtx,
-        section: 'scenario',
-      );
-      if (_activeGroup == null && mesExampleBlock.isNotEmpty) {
-        mesExampleBlock = _macroResolver.resolve(
-          mesExampleBlock,
-          macroCtx,
-          section: 'mesExample',
-        );
-      }
-      if (postHistoryBlock.isNotEmpty) {
-        postHistoryBlock = _macroResolver.resolve(
-          postHistoryBlock,
-          macroCtx,
-          section: 'postHistory',
-        );
-      }
-
-      // Impersonate instruction — comprehensive guidance for writing as the user
-      final impersonateInstruction =
-          '[System: You are now writing as $userName (the user), NOT as ${speakingCharacter.name} or any other character. '
-          'Compose $userName\'s next message in first person. '
-          'Match $userName\'s established voice, personality, and writing style from the conversation so far. '
-          'Write only $userName\'s words and actions — never narrate for ${speakingCharacter.name} or other characters. '
-          'Do not include meta-commentary, stage directions for others, or break the fourth wall. '
-          'Keep the response natural, and consistent with the scene.]\n';
-
-      // ── Context Shift: budget-aware history trimming ──
-      final fixedContent =
-          "$systemPrompt\n"
-          "$loreContent"
-          "$personaBlock\n"
-          "$userPersonaBlock"
-          "Scenario: $scenario\n"
-          "$mesExampleBlock"
-          "<START>\n"
-          "$postHistoryBlock"
-          "$authorNoteBlock"
-          "$impersonateInstruction"
-          "$suffix";
-      final fixedTokens = await _countTokens(fixedContent);
-      final contextBudget = _sessionGenSettings.resolveContextSize(
-        _storageService,
-      );
-      final generationReserve =
-          _sessionGenSettings.resolveMaxLength(_storageService) + 50;
-      final historyBudget = contextBudget - fixedTokens - generationReserve;
-
-      if (historyBudget > 0) {
-        final result = await _buildChatHistoryWithBudget(historyBudget);
-        history = result.history;
-      } else if (_messages.isNotEmpty) {
-        final lastMsg = _messages.last;
-        history = lastMsg.characterId == '__director__'
-            ? '[Director: ${lastMsg.text}]'
-            : '${lastMsg.sender}: ${lastMsg.text}';
-      }
-
-      // For chat APIs (OpenRouter, LM Studio), separate the system prompt
-      // so it can be sent as a proper 'system' role message.
-      final isRemoteApi = _llmProvider != null && !_llmProvider!.isLocal;
-      final chatSystemPrompt = isRemoteApi
-          ? "$systemPrompt\n$loreContent$personaBlock\n$userPersonaBlock"
-                "Scenario: $scenario\n$mesExampleBlock"
-          : null;
-
-      final prompt = isRemoteApi
-          ? "<START>\n"
-                "$history"
-                "$postHistoryBlock"
-                "$authorNoteBlock"
-                "$impersonateInstruction"
-                "$suffix"
-          : "$systemPrompt\n"
-                "$loreContent"
-                "$personaBlock\n"
-                "$userPersonaBlock"
-                "Scenario: $scenario\n"
-                "$mesExampleBlock"
-                "<START>\n"
-                "$history"
-                "$postHistoryBlock"
-                "$authorNoteBlock"
-                "$impersonateInstruction"
-                "$suffix";
-
-      // Stop sequences: character names only (not user — we ARE the user)
-      final g = _sessionGenSettings;
-      final stopSequences = {
-        ...g.resolveStopSequences(_storageService).toSet(),
-      };
-      if (_activeGroup != null) {
-        for (final ch in _groupCharacters) {
-          stopSequences.add('\n${ch.name}:');
-        }
-      } else {
-        stopSequences.add('\n${_activeCharacter!.name}:');
-      }
-
-      final llmService =
-          testLlmServiceOverride ??
-          _llmProvider?.activeService ??
-          _koboldService;
-      final genParams = GenerationParams(
-        prompt: prompt,
-        systemPrompt: chatSystemPrompt,
-        maxLength: g.resolveMaxLength(_storageService),
-        minLength: g.resolveMinLength(_storageService),
-        minP: g.resolveMinP(_storageService),
-        temperature: g.resolveTemperature(_storageService),
-        repeatPenalty: g.resolveRepeatPenalty(_storageService),
-        repPenTokens: g.resolveRepeatPenaltyTokens(_storageService),
-        dynatempRange: g.resolveDynamicTempEnabled(_storageService)
-            ? g.resolveDynamicTempRange(_storageService)
-            : null,
-        xtcThreshold: g.resolveXtcThreshold(_storageService),
-        xtcProbability: g.resolveXtcProbability(_storageService),
-        stopSequences: stopSequences.toList(),
-        reasoningEnabled: false,
-        reasoningEffort: g.resolveReasoningEffort(_storageService),
-        bannedPhrases: g.resolveBannedPhrases(_storageService).isNotEmpty
-            ? g.resolveBannedPhrases(_storageService)
-            : null,
-      );
-
-      final stream = llmService.generateStream(genParams);
-      String accumulated = prefix;
-      bool inThinkBlock = false;
-
-      await for (final token in stream) {
-        if (_cancelRequested) break;
-        // Filter out <think>...</think> reasoning blocks entirely
-        if (token.contains('<think>')) {
-          inThinkBlock = true;
-          continue;
-        }
-        if (token.contains('</think>')) {
-          inThinkBlock = false;
-          continue;
-        }
-        if (inThinkBlock) continue;
-        accumulated += token;
-        onToken(accumulated);
-      }
-    } catch (e) {
-      print('Impersonate error: $e');
-    } finally {
-      _isGenerating = false;
-      notifyListeners();
-    }
-  }
 
   /// Trigger the next character to speak in group mode.
   Future<void> triggerNextCharacter() async {
@@ -4867,281 +4500,12 @@ class ChatService extends ChangeNotifier {
     ]);
   }
 
-  /// Runs targeted realism evaluation for the specific character who is about
-  /// to speak next in a group chat. This is the core of making realism work
-  /// on a per-character, turn-timed basis.
-  ///
-  /// Uses temporary impersonation of _activeCharacter so that all existing
-  /// realism eval methods (_evaluateOneShotCall, _evaluateRelationshipCall, etc.)
-  /// and their parsing/inertia logic are reused without duplication.
-  Future<void> _evaluateRealismForUpcomingSpeaker(
-    CharacterCard speaker,
-  ) async {
-    // Unified gate: runs for the 1:1 host AND each group speaker (one at a time);
-    // skips group observer mode and realism-off. This is the single realism eval
-    // path — the former centralized 1:1 block was removed in favour of this.
-    if (!_realismActiveThisMode) return;
-
-    final charId = _getCharacterIdFromCard(speaker);
-    if (charId.isEmpty) return;
-
-    debugPrint(
-      '[Realism:Unified] Pre-turn eval for upcoming speaker: ${speaker.name} '
-      '($charId) — mode=${_activeGroup == null ? "1:1" : "group"}',
-    );
-
-    // Save previous 1:1 context (normally null in pure group sessions)
-    final previousActiveCharacter = _activeCharacter;
-
-    // Impersonate this speaker for the duration of the eval so all existing
-    // LLM eval methods, guards, name/personality reads, and delta application
-    // logic work exactly as they do for 1:1 chats.
-    _activeCharacter = speaker;
-
-    // Group non-observer: ensure this definite speaker receives their per-turn needs decay
-    // (central tick in sendMessage is skipped for groups to support random turn order without
-    // always decaying the 'first' member). Snapshot the pre-decay value for chips/realism_state
-    // *before* applying this turn's decay, then decay the speaker's map entry, then load scalars.
-    if (_activeGroup != null && !_observerMode && _needsSimEnabled) {
-      final sidForDecay = charId;
-      final currentForSpeaker = _getGroupNeeds(sidForDecay);
-      final preDecay = currentForSpeaker.isNotEmpty
-          ? Map<String, int>.from(currentForSpeaker)
-          : {
-              for (final k in NeedsSimulation.needKeys)
-                k: NeedsSimulation.needDefaults[k] ?? 80,
-            };
-      // Stash the true pre-decay for this speaker so post-gen chip delta computation
-      // (and regen) see the correct baseline including the decay portion of the turn.
-      _pendingRealismMetadata ??= {};
-      _pendingRealismMetadata!['needs_pre_turn_vector'] = preDecay;
-
-      // Apply one tick of decay directly to this speaker's group entry (custom rates or defaults).
-      final decayed = Map<String, int>.from(preDecay);
-      final customRates = _groupDecayRates;
-      for (final key in NeedsSimulation.needKeys) {
-        final cur = decayed[key] ?? 80;
-        final decay = customRates[key] ?? NeedsSimulation.needDecay[key] ?? 0;
-        decayed[key] = (cur - decay).clamp(0, 100);
-      }
-      _setGroupNeeds(sidForDecay, decayed);
-
-      // Now load the post-decay state into scalars for the remainder of the speaker eval + prompt injection.
-      _loadGroupRealismIntoScalars(charId);
-    } else {
-      // Load this speaker's persisted group realism state into the scalar fields
-      // that the eval methods will read and mutate.
-      _loadGroupRealismIntoScalars(charId);
-    }
-
-    // Phase 2: Ensure hidden inter-character relationship tracking is seeded
-    // for all other group members (neutral 0). This happens on the speaker's
-    // first turn with realism so the invisible feelings map is always present.
-    _relationshipService.ensureInterCharacterRelationshipsSeeded(charId);
-
-    _isEvaluatingRealism = true;
-    _realismEvalStreamText = '';
-    notifyListeners();
-
-    // Capture this speaker's pre-turn needs vector (before decay + eval)
-    Map<String, int>? preTurnVector;
-    if (_needsSimEnabled && _needsSimulation.vector.isNotEmpty) {
-      preTurnVector = Map<String, int>.from(_needsSimulation.vector);
-    }
-
-    // Temporarily load this speaker's personal objectives so the narrative
-    // evaluation (and one-shot) sees the correct primary/secondary context
-    // for "proposed_objective" generation. This is required for 1:1 parity.
-    final previousObjectives = List<Objective>.from(_activeObjectives);
-    final speakerObjectives = await getActiveObjectivesFor(speaker);
-    _activeObjectives = speakerObjectives.where((o) => o.active).toList();
-
-    void handleChunk(String chunk) {
-      _realismEvalStreamText += chunk;
-      _evalChunkTimer?.cancel();
-      _evalChunkTimer = Timer(const Duration(milliseconds: 150), () {
-        try {
-          notifyListeners();
-        } catch (_) {}
-      });
-    }
-
-    try {
-      // Respect early cancellation
-      if (_realismEvalCancelled) {
-        debugPrint(
-          '[Realism:Group] Evaluation cancelled before LLM calls for ${speaker.name}',
-        );
-        _realismEvalCancelled = false;
-        return;
-      }
-
-      if (_relationshipService.pendingTrustRepair) {
-        // Trust-repair eval (fires when trust dropped sharply). Was 1:1-only in
-        // the old centralized block; now part of the single path so the host
-        // keeps it (and a group member would too, if their trust ever flags it).
-        debugPrint(
-          '[Realism:Unified] Trust-repair eval for ${speaker.name} ($charId)',
-        );
-        _relationshipService.consumePendingTrustRepair();
-        final userText = _messages
-            .lastWhere(
-              (m) => m.isUser,
-              orElse: () => ChatMessage(text: '', sender: '', isUser: true),
-            )
-            .text;
-        await _evaluateTrustRepairCall(userText, onChunk: handleChunk);
-      } else if (_storageService.realismSettings.realismOneShotEval) {
-        debugPrint(
-          '[Realism:Unified] One-shot eval for ${speaker.name} ($charId)',
-        );
-        await _evaluateOneShotCall(onChunk: handleChunk);
-      } else {
-        // Run the four evals AND the batched verifier pass — identical to the
-        // (former) centralized 1:1 path, so EVERY speaker (host or group member)
-        // gets the same double-checked realism. This is the parity unification.
-        debugPrint(
-          '[Realism:Unified] 4-call eval + verifier for ${speaker.name} ($charId)',
-        );
-        _realismEvals.beginCollectForBatchedVerification();
-        await _fireStaggeredRealismEvals(handleChunk);
-        await _realismEvals.finalizeBatchedRealismVerifications();
-
-        final collected = _realismEvals.getCollectedForBatch();
-        if (collected.isNotEmpty) {
-          debugPrint(
-            '[Realism:Unified] Verifying ${collected.length} eval(s) for '
-            '${speaker.name}',
-          );
-          final items = collected
-              .map(
-                (p) => (
-                  evalKind: p['kind'] as String,
-                  rawOutput: p['raw'] as String,
-                  sceneResponse: p['scene'] as String,
-                  preState: null,
-                  activeChar: _activeCharacter,
-                  activeGroup: _activeGroup,
-                  recentMessages: _messages,
-                  promptText: p['prompt'] as String?,
-                  injections: (p['injections'] as Map?)?.cast<String, String>(),
-                  strictnessOverride: null,
-                  maxPassesOverride: null,
-                ),
-              )
-              .toList();
-          final batchRes = await _realismVerifier.verifyBatch(items);
-          await _realismEvals.applyBatchResults(batchRes);
-        }
-      }
-
-      // Handle cancellation after the eval calls
-      if (_realismEvalCancelled) {
-        debugPrint(
-          '[Realism:Unified] Evaluation cancelled during/after LLM calls for ${speaker.name}',
-        );
-        _realismEvalCancelled = false;
-        return;
-      }
-
-      // Harvest the now-updated scalar fields back into this speaker's
-      // _groupRealism entry so prompt injection and UI see fresh values.
-      _saveScalarsIntoGroupRealism(charId);
-
-      // Synthesize metadata for timeline / chips (best-effort, same as 1:1 path)
-      _pendingRealismMetadata ??= {};
-      _pendingRealismMetadata!['emotion_label'] = _characterEmotion;
-      _pendingRealismMetadata!['realism_state'] = _captureRealismState(
-        preTurn: preTurnVector,
-      );
-
-      if (_needsSimEnabled) {
-        final needsDeltas = _needsSimulation.computeNeedsDeltasWithReasons(
-          preTurnVector ?? const <String, int>{},
-        );
-        if (needsDeltas.isNotEmpty) {
-          _pendingRealismMetadata!['needs_deltas'] = needsDeltas;
-        }
-      }
-
-      _saveChat();
-    } finally {
-      // Always restore previous context and clear busy state
-      _activeCharacter = previousActiveCharacter;
-      _activeObjectives = previousObjectives;
-      _evalChunkTimer?.cancel();
-      _evalChunkTimer = null;
-      _isEvaluatingRealism = false;
-      notifyListeners();
-    }
-  }
-
-  /// Loads the given group character's realism values from _groupRealism into
-  /// the single-character scalar fields so the existing eval methods can
-  /// operate on them during impersonation.
-  void _loadGroupRealismIntoScalars(String charId) {
-    // Relationship (affection/trust/fix/tiers etc) now via service load helper (uses the same _getGroup* internally via cbs).
-    _relationshipService.loadRelationshipScalarsForSpeaker(charId);
-    // Nsfw (arousal + cooldown + nsfwEnabled per char) via service (extends prior arousal-only for full group parity).
-    // Note: group uses 'arousal' key (historical) vs snapshot 'arousalLevel' for compat.
-    _nsfwService.loadNsfwScalarsForSpeaker(charId);
-
-    _characterEmotion = _getGroupString(charId, 'emotion');
-    _emotionIntensity = _getGroupString(
-      charId,
-      'emotionIntensity',
-      defaultValue: 'moderate',
-    );
-
-    // Needs vector (if any persisted for this char)
-    final needs = _getGroupNeeds(charId);
-    if (needs.isNotEmpty) {
-      _needsSimulation.restoreFromSnapshot({'vector': needs});
-    } else if (_needsSimEnabled) {
-      // Fresh start for a group member who has never had needs for this group chat.
-      // Use full 100 to match 1:1 "new chat" behavior (prevents bleed perception).
-      _needsSimulation.initializeFresh();
-    }
-  }
-
-  /// Writes the current scalar realism fields back into the target group
-  /// character's _groupRealism entry after an impersonated eval round.
-  void _saveScalarsIntoGroupRealism(String charId) {
-    // Relationship scalars (affection/long/trust/fix/tiers/spatial) now via service.
-    _relationshipService.saveRelationshipScalarsToGroup(charId);
-    // Nsfw scalars (arousal + cooldown + enabled) now via service (for group per-char persistence parity).
-    // Note: group uses 'arousal' key (historical) vs snapshot 'arousalLevel' for compat.
-    _nsfwService.saveNsfwScalarsToGroup(charId);
-
-    if (_characterEmotion.isNotEmpty) {
-      _setGroupRealismValue(charId, 'emotion', _characterEmotion);
-    }
-    if (_emotionIntensity.isNotEmpty) {
-      _setGroupRealismValue(charId, 'emotionIntensity', _emotionIntensity);
-    }
-
-    // Persist current needs vector for this speaker
-    if (_needsSimulation.vector.isNotEmpty) {
-      _setGroupNeeds(charId, Map<String, int>.from(_needsSimulation.vector));
-    }
-  }
-
-  /// Public API: Focus the personal objectives of a specific group member so the
-  /// existing objective management UI and generation can operate on them.
-  /// Does nothing in 1:1 mode.
-  Future<void> focusObjectivesForGroupCharacter(CharacterCard character) async {
-    if (_activeGroup == null) return;
-    final charId = _getCharacterIdFromCard(character);
-    final objs = await _db.getObjectivesForCharacter(
-      charId,
-      chatId: _currentSessionId,
-    );
-    _activeObjectives = objs.where((o) => o.active).toList();
-    notifyListeners();
-  }
-
   /// Loads the active objectives for the given character in the current session.
   /// Safe to call from group objective UIs — does not mutate global _activeObjectives.
+  ///
+  /// Kept in the class body (not an extension) because [FakeChatService]
+  /// overrides it in golden tests — extension members are statically dispatched
+  /// and cannot be overridden.
   Future<List<Objective>> getActiveObjectivesFor(
     CharacterCard character,
   ) async {
@@ -5153,231 +4517,6 @@ class ChatService extends ChangeNotifier {
       debugPrint('[Objective] Failed to load for ${character.name}: $e');
       return const [];
     }
-  }
-
-  // ── Group Creation Baseline Seeding (bond/trust/emotion/time/day only) ──
-
-  /// Returns the immutable creation-time baseline realism values for a group member.
-  /// Only the allowed seeding fields are exposed: affection (bond), trust, emotion, timeOfDay, dayCount.
-  Map<String, dynamic> getBaselineSeedForGroupCharacter(
-    CharacterCard character,
-  ) {
-    if (_activeGroup == null) return {};
-    final charId = _getCharacterIdFromCard(character);
-    try {
-      final json = jsonDecode(_activeGroup!.baselineRealismState);
-      if (json is Map && json.containsKey(charId)) {
-        final data = json[charId] as Map<String, dynamic>? ?? {};
-        return {
-          'affection': (data['affection'] as num?)?.toInt() ?? 50,
-          'trust': (data['trust'] as num?)?.toInt() ?? 50,
-          'emotion': (data['emotion'] as String?) ?? 'neutral',
-          'emotionIntensity':
-              (data['emotionIntensity'] as String?) ?? 'moderate',
-          'timeOfDay': (data['timeOfDay'] as String?) ?? 'morning',
-          'dayCount': (data['dayCount'] as num?)?.toInt() ?? 1,
-        };
-      }
-    } catch (_) {}
-    return {
-      'affection': 50,
-      'trust': 50,
-      'emotion': 'neutral',
-      'emotionIntensity': 'moderate',
-      'timeOfDay': 'morning',
-      'dayCount': 1,
-    };
-  }
-
-  /// Updates the immutable creation baseline for a group member.
-  /// Only allowed fields are accepted. This should only be called during group creation seeding.
-  void setBaselineSeedForGroupCharacter(
-    CharacterCard character,
-    Map<String, dynamic> values,
-  ) {
-    if (_activeGroup == null) return;
-    final charId = _getCharacterIdFromCard(character);
-
-    Map<String, dynamic> baseline;
-    try {
-      baseline = Map<String, dynamic>.from(
-        jsonDecode(_activeGroup!.baselineRealismState),
-      );
-    } catch (_) {
-      baseline = {};
-    }
-
-    baseline[charId] = {
-      'affection': (values['affection'] as num?)?.toInt() ?? 50,
-      'trust': (values['trust'] as num?)?.toInt() ?? 50,
-      'emotion': (values['emotion'] as String?) ?? 'neutral',
-      'emotionIntensity': (values['emotionIntensity'] as String?) ?? 'moderate',
-      'timeOfDay': (values['timeOfDay'] as String?) ?? 'morning',
-      'dayCount': (values['dayCount'] as num?)?.toInt() ?? 1,
-    };
-
-    _activeGroup!.baselineRealismState = jsonEncode(baseline);
-    notifyListeners();
-  }
-
-  /// Loads the personal objectives for the current/next speaker into _activeObjectives
-  /// when in group mode. This makes the existing objective UI, generation, and injection
-  /// work per-character in groups without duplicating the entire objective system.
-  Future<void> _loadObjectivesForCurrentSpeaker() async {
-    if (_activeGroup == null || _currentSessionId == null) return;
-
-    final speaker = nextCharacter ?? _groupCharacters.firstOrNull;
-    if (speaker == null) {
-      _activeObjectives = [];
-      _messagesSinceLastCheck = 0;
-      _isCheckingCompletion = false;
-      _summaryPaused =
-          false; // explicit secondary zero for _summaryPaused (symmetric; _loadObjectivesForCurrentSpeaker no-speaker hygiene)
-      _isSummaryGenerating =
-          false; // secondary zero in _loadObjectivesForCurrentSpeaker no-speaker (group hygiene for summary flag)
-      _userMessagesSinceLastPeriodicEval = 0;
-      _isExtractingFacts =
-          false; // secondary fact flag + counter zero in _loadObjectivesForCurrentSpeaker no-speaker (group hygiene; fact_extraction)
-      _isEvolvingCharacter = false;
-      _evolutionStatus = '';
-      _evolutionError =
-          ''; // explicit evo flag/status/error zero in _loadObjectivesForCurrentSpeaker no-speaker (group hygiene; evolution_service (stateless or prompt-only; no reset calls needed))
-      notifyListeners();
-      return;
-    }
-
-    final charId = _getCharacterIdFromCard(speaker);
-    final objs = await _db.getObjectivesForCharacter(
-      charId,
-      chatId: _currentSessionId,
-    );
-
-    _activeObjectives = objs.where((o) => o.active).toList();
-    notifyListeners();
-  }
-
-  /// One-time seeding of objectives that were carried in an imported Group Card.
-  /// Called after group state is loaded for a freshly imported group.
-  Future<void> _seedImportedMemberObjectivesIfPresent() async {
-    if (_activeGroup == null || _currentSessionId == null) return;
-
-    try {
-      final stateJson = _activeGroup!.defaultMemberRealismState;
-      if (stateJson.isEmpty || stateJson == '{}') return;
-
-      final map = jsonDecode(stateJson);
-      if (map is! Map) return;
-
-      final importedObj = map['imported_member_objectives'];
-      if (importedObj is! Map) return;
-
-      for (final entry in importedObj.entries) {
-        final charId = entry.key.toString();
-        final list = entry.value as List? ?? [];
-        for (final objData in list) {
-          final objMap = objData as Map<String, dynamic>? ?? {};
-          final newId =
-              'obj_${DateTime.now().millisecondsSinceEpoch}_${charId.hashCode}';
-          await _db.insertObjective(
-            ObjectivesCompanion.insert(
-              id: newId,
-              characterId: charId,
-              chatId: drift.Value(_currentSessionId!),
-              objective:
-                  objMap['objective']?.toString() ?? 'Imported objective',
-              tasks: drift.Value(objMap['tasks']?.toString() ?? '[]'),
-              active: const drift.Value(true),
-              isPrimary: drift.Value(objMap['isPrimary'] == true),
-              checkFrequency: drift.Value(
-                (objMap['checkFrequency'] as num?)?.toInt() ?? 3,
-              ),
-              injectionDepth: drift.Value(
-                (objMap['injectionDepth'] as num?)?.toInt() ?? 4,
-              ),
-            ),
-          );
-        }
-      }
-
-      // Remove the marker so it doesn't seed again
-      map.remove('imported_member_objectives');
-      _activeGroup!.defaultMemberRealismState = jsonEncode(map);
-      await _saveChat();
-    } catch (_) {}
-  }
-
-  String _getRealismStateInjection() {
-    // Thin delegation to the new central realism state composer.
-    // This is the single source of the grouped "Speaker Internal State" block
-    // that the model receives (metrics first + guidance). Replaces the old
-    // manual 8-builder concat.
-    return _realismStateInjection.buildRealismStateInjection();
-  }
-
-  void _restoreRealismStateFromMessage(ChatMessage? msg) {
-    if (msg == null) return;
-
-    // Check if the current visible node has an active swipe metadata array or just the base metadata
-    final meta = msg.activeMetadata;
-    if (meta == null || !meta.containsKey('realism_state')) {
-      debugPrint(
-        '[Realism] No time-travel snapshot found in message. Legacy state kept.',
-      );
-      return;
-    }
-
-    final state = meta['realism_state'] as Map<String, dynamic>;
-    _relationshipService.restoreFromMessageState(state);
-    _moodDecayCounter = state['moodDecayCounter'] as int? ?? _moodDecayCounter;
-    _characterEmotion =
-        state['characterEmotion'] as String? ?? _characterEmotion;
-    _emotionIntensity =
-        state['emotionIntensity'] as String? ?? _emotionIntensity;
-
-    _timeService.restoreTimeFromRealismState(state);
-
-    _nsfwService.restoreNsfwFromRealismState(state);
-
-    // v3.0 Restorations (relationship via service; already covered by restoreFromMessageState above for most).
-    // (Direct sets removed; service owns the scalars.)
-
-    // Needs simulation snapshot (clean port)
-    // Only restore the vector if the sim is currently enabled for this session.
-    // Never let a historical snapshot flip _needsSimEnabled back on (supports
-    // clean mid-chat toggle-off via setNeedsSimEnabled without stale state).
-    if (state.containsKey('needs') &&
-        state['needs'] is Map &&
-        _needsSimEnabled) {
-      final needsData = state['needs'] as Map;
-      _needsSimulation.restoreFromSnapshot(needsData);
-    }
-
-    debugPrint(
-      '[Realism] Engine state successfully rolled back to match timeline.',
-    );
-  }
-
-  /// Runs all post-generation needs-related checks (climax, sexual activity,
-  /// daily activities, fulfillment) via thin delegate to the consolidated
-  /// NeedsImpactEvaluator (simple model + optional Director authority review loop).
-  /// Orchestration (guards, group impersonation dance + loadGroupRealismIntoScalars
-  /// before call so prompts see correct $charName/personality/stance, preTurn
-  /// snapshot for chips, post _saveScalarsIntoGroupRealism + attach needs_deltas,
-  /// (orchestration + impersonation dance in god; full in evaluator).
-  Future<void> _runPostGenNeedsChecks(String responseText) async {
-    await _needsImpactEvaluator.evaluateAndApply(responseText);
-  }
-
-  // (unified thin + evaluator; prior _check* excised as dead. See CLAUDE.md).
-
-  // ── Score / State Helpers (thinned; core logic + counters in RelationshipService) ──
-
-  /// Apply short-term relationship decay (2 points per 10 turns toward 0)
-  /// This prevents relationships from being permanently stuck at extremes.
-  void _applyMoodDecay() {
-    // Decay mechanism moved to RelationshipService (applyShortTermDecay).
-    // Counter, 1:1/group branches, inter-char decay all delegated for mechanical fidelity.
-    _relationshipService.applyShortTermDecay();
   }
 
   // ── Public Toggle Methods ──
